refactor: use unifiedjs esm packages
