+++ lib/remark-code-title.js
@@ -1,6 +1,6 @@
-import visit from 'unist-util-visit'
+import { visit } from 'unist-util-visit'
 
-module.exports = function (options) {
+export default function remarkCodeTitles() {
   return (tree) =>
     visit(tree, 'code', (node, index) => {
       const nodeLang = node.lang || ''
